+++ packages/frameworks/react/CHANGELOG.md
@@ -6,6 +6,8 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
+## [2.0.2] - 2024-02-10
+
 ### Added
 
 - Exported `SelectionDetails` type for `Menu` component